refactor: update auth form section schema imports and remove deprecated file
- Changed import paths for AuthFormSectionSchema to use the new naming convention.
- Deleted the old AuthFormSectionSchema file to streamline the codebase and avoid confusion.
