+++ src/shared/hooks/useDateTime.ts
@@ -9,14 +9,14 @@ import { formatMonthDay, formatTime } from "@/src/shared/utils/timezone";
 export function useFormattedTime(
   format: "short" | "long" = "long",
 ): string | null {
-  const [formattedTime, setFormattedTime] = useState<string | null>(null);
+  const [formattedTime, setFormattedTime] = useState<string | null>(() =>
+    formatTime(new Date(), format),
+  );
 
   useEffect(() => {
-    // 初回実行
     const updateTime = () => {
       setFormattedTime(formatTime(new Date(), format));
     };
-    updateTime();
 
     // 1分ごとに更新
     const intervalId = setInterval(updateTime, 60000);
@@ -30,11 +30,11 @@ export function useFormattedTime(
 
 /**
  * リアルタイムで月日をフォーマットして返すフック
- * 1分ごとに自動更新される（日付が変わった場合に対応）
+ * 10分ごとに自動更新される（日付が変わった場合に対応）
  */
 export function useFormattedMonthDay(): string | null {
   const [formattedMonthDay, setFormattedMonthDay] = useState<string | null>(
-    null,
+    formatMonthDay(new Date()),
   );
 
   useEffect(() => {
